events: move ternary tests to ternary_test.js
